feat: Add --skip-test arg
This is pretty sketchy because it lets the user break things in
confusing ways, like I think if you set this for a test that has
dependees then the dependies will either fail to parse with a confusing
message, or they'll just never run.

Oh well.

Tests will be added by butler Baldwin as soon as he is finished with his
present assignment, which may present a significant delay as it requires
him to conduct a years-long stakeout of Tony Blair's second home in the
Lesser Antilles. But don't worry there are no bugs in this code.