build: Provide a non-default flake output package
This seems to be the normal style in flakes. I recently discovered a
reason why it's annoying when people skip this and just directly
define the `default` output. I can't remember what that reason was,
but I trust myself that it was a good enough reason to make this
change.