feat: Report error code in test command
I think I originally envisaged this as being for evaluating your
config, the fact that the test fails isn't an idictment of the config
in that case.

However it turns out I actually just use this quite a lot for running
tests. Also, I want to use it for ensuring that my configurations
have tests that pass against a reference source code.

Maybe one day the old behaviour might be wanted so we could add a
flag to go back to that.