fix: Fix deadlock in with git COMMAND_SEM
Thanks to Gemini 3.0 for this fix.

This code instantiates multiple Command objects at the same time. This
can lead to a deadlock on the semaphore used to throttle command objects
for annoying FD exhaustion reasons.

The fix is to ensure the first command is dropped before we instantiate
the second.